lec_chathelper/requirements.txt
asgiref==3.6.0
bitstring==4.0.1
cffi==1.15.1
Django==4.1.7
django-environ==0.10.0
install==1.3.5
psycopg2-binary==2.9.5
pycparser==2.21
pycryptodome-test-vectors==1.0.12
pycryptodomex==3.17
pyjwt==2.6.0
sqlparse==0.4.3
boto3==1.26.99
djangorestframework==3.14.0
botocore==1.29.100
redis==4.5.4
uuid==1.30
pyyaml==6.0
celery==5.2.7
kombu==5.2.4
confluent-kafka==1.9.2
django-rest-swagger==2.2.0
uritemplate==4.1.1


# ASR app
faster-whisper==0.4.1
torch
openai-whisper==20230124